Added the EVAL of the base-class-expander back to def-view-class/meta
This does not work as a macro as the information it needs is not available at compile time. I've punted to EVAL for now (EVAL'ing a defclass form is not so bad), but a better, most probably MOP based solution will need to be found.

darcs-hash:20050704073748-5417e-79ea36affdf24647cfe095d90c9c378491822fc3.gz